+++ lib/hachoir/field/byte_field.py
@@ -3,8 +3,6 @@ Very basic field: raw content with a size in byte. Use this class for
 unknown content.
 """
 
-import types
-
 from hachoir.field import Field, FieldError
 from hachoir.core.tools import makePrintable
 from hachoir.core.bits import str2hex
@@ -30,9 +28,9 @@ class RawBytes(Field):
 
     def _createDisplay(self, human):
         max_bytes = config.max_byte_length
-        if isinstance(self._getValue, types.FunctionType):
+        try:
             display = makePrintable(self.value[:max_bytes], "ASCII")
-        else:
+        except Exception:
             if self._display is None:
                 address = self.absolute_address
                 length = min(self._size // 8, max_bytes)