+++ lib/hachoir/parser/program/java_serialized.py
@@ -138,10 +138,16 @@ class ObjectValue(FieldSet):
             classDesc = classDesc.referent
         if isinstance(classDesc, SerializedNull):
             return
-        # TODO: proxy class desc
 
         for field in self.gen_values(classDesc['superClassDesc']):
             yield field
+
+        if isinstance(classDesc, SerializedProxyClassDesc):
+            return
+
+        if classDesc['classDescFlags_externalizable'].value:
+            yield WriteObjectContents(self, "external[]", "%s.writeExternal() output" % classDesc['className'].value)
+            return
 
         for fieldDesc in classDesc.array('fieldDesc'):
             tc = fieldDesc['typecode'].value
@@ -275,7 +281,17 @@ class SerializedLongString(FieldSet):
 class SerializedProxyClassDesc(FieldSet):
     def createFields(self):
         yield Enum(UInt8(self, "typecode"), TYPECODE_NAMES)
-        # TODO
+
+        self.root.newHandle(self)
+        yield Int32(self, "proxyInterfaceName_count")
+        for i in range(self['proxyInterfaceName_count'].value):
+            yield PascalString16(self, "proxyInterfaceName[]", charset="UTF-8")
+        yield ClassAnnotation(self, "classAnnotation")
+        yield SerializedContent(self, "superClassDesc")
+
+    @property
+    def className(self):
+        return '<Proxy implements %s>' % (', '.join(v.value for v in self.array('proxyInterfaceName')))
 
 
 class SerializedEnum(FieldSet):