+++ CHANGES.md
@@ -7,6 +7,7 @@
 * Update Certifi 2018.11.29 (10a1f8a) to 2019.03.09 (401100f)
 * Update dateutil 2.7.5 (e954819) to 2.8.0 (c90a30c)
 * Update DiskCache library 3.1.1 (05cac6a) to 3.1.1 (2649ac9)
+* Update Hachoir library 2.0a3 to 2.0a6 (c102cc7)
 
 
 [develop changelog]